agents/assistant/requirements.txt
google-adk[a2a]
mcp
pyyaml
uvicorn
